Starting to getting getFacts() method out of IncrementalIndex
