+++ processing/src/main/java/io/druid/segment/incremental/IncrementalIndex.java
@@ -25,7 +25,6 @@ import com.google.common.base.Strings;
 import com.google.common.base.Supplier;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Iterators;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.primitives.Ints;
@@ -34,7 +33,6 @@ import io.druid.collections.NonBlockingPool;
 import io.druid.common.config.NullHandling;
 import io.druid.common.guava.GuavaUtils;
 import io.druid.data.input.InputRow;
-import io.druid.data.input.MapBasedRow;
 import io.druid.data.input.Row;
 import io.druid.data.input.impl.DimensionSchema;
 import io.druid.data.input.impl.DimensionsSpec;
@@ -83,16 +81,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.Deque;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentLinkedDeque;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.ConcurrentNavigableMap;
-import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -227,7 +219,7 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
   private final boolean rollup;
   private final List<Function<InputRow, InputRow>> rowTransformers;
   private final VirtualColumns virtualColumns;
-  private final AggregatorFactory[] metrics;
+  protected final AggregatorFactory[] metrics;
   private final AggregatorType[] aggs;
   private final boolean deserializeComplexMetrics;
   private final boolean reportParseExceptions;
@@ -437,8 +429,6 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
     return rollup;
   }
 
-  protected abstract FactsHolder getFacts();
-
   public abstract boolean canAppendRow();
 
   public abstract String getOutOfRowsReason();
@@ -479,6 +469,7 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
 
   protected abstract boolean isNull(int rowOffset, int aggOffset);
 
+  public abstract Iterable<TimeAndDims> timeRangeIterable(boolean descending, long timeStart, long timeEnd);
 
   @Override
   public void close()
@@ -831,57 +822,7 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
     return iterableWithPostAggregations(null, false).iterator();
   }
 
-  public Iterable<Row> iterableWithPostAggregations(final List<PostAggregator> postAggs, final boolean descending)
-  {
-    return new Iterable<Row>()
-    {
-      @Override
-      public Iterator<Row> iterator()
-      {
-        final List<DimensionDesc> dimensions = getDimensions();
-
-        return Iterators.transform(
-            getFacts().iterator(descending),
-            timeAndDims -> {
-              final int rowOffset = timeAndDims.getRowIndex();
-
-              Object[] theDims = timeAndDims.getDims();
-
-              Map<String, Object> theVals = Maps.newLinkedHashMap();
-              for (int i = 0; i < theDims.length; ++i) {
-                Object dim = theDims[i];
-                DimensionDesc dimensionDesc = dimensions.get(i);
-                if (dimensionDesc == null) {
-                  continue;
-                }
-                String dimensionName = dimensionDesc.getName();
-                DimensionHandler handler = dimensionDesc.getHandler();
-                if (dim == null || handler.getLengthOfEncodedKeyComponent(dim) == 0) {
-                  theVals.put(dimensionName, null);
-                  continue;
-                }
-                final DimensionIndexer indexer = dimensionDesc.getIndexer();
-                Object rowVals = indexer.convertUnsortedEncodedKeyComponentToActualArrayOrList(dim, DimensionIndexer.LIST);
-                theVals.put(dimensionName, rowVals);
-              }
-
-              AggregatorType[] aggs = getAggsForRow(rowOffset);
-              for (int i = 0; i < aggs.length; ++i) {
-                theVals.put(metrics[i].getName(), getAggVal(aggs[i], rowOffset, i));
-              }
-
-              if (postAggs != null) {
-                for (PostAggregator postAgg : postAggs) {
-                  theVals.put(postAgg.getName(), postAgg.compute(theVals));
-                }
-              }
-
-              return new MapBasedRow(timeAndDims.getTimestamp(), theVals);
-            }
-        );
-      }
-    };
-  }
+  public abstract Iterable<Row> iterableWithPostAggregations(List<PostAggregator> postAggs, boolean descending);
 
   public DateTime getMaxIngestedEventTime()
   {
@@ -1034,7 +975,7 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
       return rowIndex;
     }
 
-    private void setRowIndex(int rowIndex)
+    public void setRowIndex(int rowIndex)
     {
       this.rowIndex = rowIndex;
     }
@@ -1171,227 +1112,6 @@ public abstract class IncrementalIndex<AggregatorType> extends AbstractIndex imp
       }
     }
     return true;
-  }
-
-  interface FactsHolder
-  {
-    /**
-     * @return the previous rowIndex associated with the specified key, or
-     * {@code TimeAndDims#EMPTY_ROW_INDEX} if there was no mapping for the key.
-     */
-    int getPriorIndex(TimeAndDims key);
-
-    long getMinTimeMillis();
-
-    long getMaxTimeMillis();
-
-    Iterator<TimeAndDims> iterator(boolean descending);
-
-    Iterable<TimeAndDims> timeRangeIterable(boolean descending, long timeStart, long timeEnd);
-
-    Iterable<TimeAndDims> keySet();
-
-    /**
-     * @return the previous rowIndex associated with the specified key, or
-     * {@code TimeAndDims#EMPTY_ROW_INDEX} if there was no mapping for the key.
-     */
-    int putIfAbsent(TimeAndDims key, int rowIndex);
-
-    void clear();
-  }
-
-  static class RollupFactsHolder implements FactsHolder
-  {
-    private final boolean sortFacts;
-    // Can't use Set because we need to be able to get from collection
-    private final ConcurrentMap<TimeAndDims, TimeAndDims> facts;
-    private final List<DimensionDesc> dimensionDescsList;
-
-    public RollupFactsHolder(boolean sortFacts, Comparator<TimeAndDims> timeAndDimsComparator, List<DimensionDesc> dimensionDescsList)
-    {
-      this.sortFacts = sortFacts;
-      if (sortFacts) {
-        this.facts = new ConcurrentSkipListMap<>(timeAndDimsComparator);
-      } else {
-        this.facts = new ConcurrentHashMap<>();
-      }
-      this.dimensionDescsList = dimensionDescsList;
-    }
-
-    @Override
-    public int getPriorIndex(TimeAndDims key)
-    {
-      TimeAndDims timeAndDims = facts.get(key);
-      return timeAndDims == null ? TimeAndDims.EMPTY_ROW_INDEX : timeAndDims.rowIndex;
-    }
-
-    @Override
-    public long getMinTimeMillis()
-    {
-      if (sortFacts) {
-        return ((ConcurrentNavigableMap<TimeAndDims, TimeAndDims>) facts).firstKey().getTimestamp();
-      } else {
-        throw new UnsupportedOperationException("can't get minTime from unsorted facts data.");
-      }
-    }
-
-    @Override
-    public long getMaxTimeMillis()
-    {
-      if (sortFacts) {
-        return ((ConcurrentNavigableMap<TimeAndDims, TimeAndDims>) facts).lastKey().getTimestamp();
-      } else {
-        throw new UnsupportedOperationException("can't get maxTime from unsorted facts data.");
-      }
-    }
-
-    @Override
-    public Iterator<TimeAndDims> iterator(boolean descending)
-    {
-      if (descending && sortFacts) {
-        return ((ConcurrentNavigableMap<TimeAndDims, TimeAndDims>) facts).descendingMap().keySet().iterator();
-      }
-      return keySet().iterator();
-    }
-
-    @Override
-    public Iterable<TimeAndDims> timeRangeIterable(boolean descending, long timeStart, long timeEnd)
-    {
-      if (!sortFacts) {
-        throw new UnsupportedOperationException("can't get timeRange from unsorted facts data.");
-      }
-      TimeAndDims start = new TimeAndDims(timeStart, new Object[]{}, dimensionDescsList);
-      TimeAndDims end = new TimeAndDims(timeEnd, new Object[]{}, dimensionDescsList);
-      ConcurrentNavigableMap<TimeAndDims, TimeAndDims> subMap =
-          ((ConcurrentNavigableMap<TimeAndDims, TimeAndDims>) facts).subMap(start, end);
-      final Map<TimeAndDims, TimeAndDims> rangeMap = descending ? subMap.descendingMap() : subMap;
-      return rangeMap.keySet();
-    }
-
-    @Override
-    public Iterable<TimeAndDims> keySet()
-    {
-      return facts.keySet();
-    }
-
-    @Override
-    public int putIfAbsent(TimeAndDims key, int rowIndex)
-    {
-      // setRowIndex() must be called before facts.putIfAbsent() for visibility of rowIndex from concurrent readers.
-      key.setRowIndex(rowIndex);
-      TimeAndDims prev = facts.putIfAbsent(key, key);
-      return prev == null ? TimeAndDims.EMPTY_ROW_INDEX : prev.rowIndex;
-    }
-
-    @Override
-    public void clear()
-    {
-      facts.clear();
-    }
-  }
-
-  static class PlainFactsHolder implements FactsHolder
-  {
-    private final boolean sortFacts;
-    private final ConcurrentMap<Long, Deque<TimeAndDims>> facts;
-
-    public PlainFactsHolder(boolean sortFacts)
-    {
-      this.sortFacts = sortFacts;
-      if (sortFacts) {
-        this.facts = new ConcurrentSkipListMap<>();
-      } else {
-        this.facts = new ConcurrentHashMap<>();
-      }
-    }
-
-    @Override
-    public int getPriorIndex(TimeAndDims key)
-    {
-      // always return EMPTY_ROW_INDEX to indicate that no prior key cause we always add new row
-      return TimeAndDims.EMPTY_ROW_INDEX;
-    }
-
-    @Override
-    public long getMinTimeMillis()
-    {
-      if (sortFacts) {
-        return ((ConcurrentNavigableMap<Long, Deque<TimeAndDims>>) facts).firstKey();
-      } else {
-        throw new UnsupportedOperationException("can't get minTime from unsorted facts data.");
-      }
-    }
-
-    @Override
-    public long getMaxTimeMillis()
-    {
-      if (sortFacts) {
-        return ((ConcurrentNavigableMap<Long, Deque<TimeAndDims>>) facts).lastKey();
-      } else {
-        throw new UnsupportedOperationException("can't get maxTime from unsorted facts data.");
-      }
-    }
-
-    @Override
-    public Iterator<TimeAndDims> iterator(boolean descending)
-    {
-      if (descending && sortFacts) {
-        return concat(((ConcurrentNavigableMap<Long, Deque<TimeAndDims>>) facts)
-                .descendingMap().values(), true).iterator();
-      }
-      return concat(facts.values(), false).iterator();
-    }
-
-    @Override
-    public Iterable<TimeAndDims> timeRangeIterable(boolean descending, long timeStart, long timeEnd)
-    {
-      ConcurrentNavigableMap<Long, Deque<TimeAndDims>> subMap =
-          ((ConcurrentNavigableMap<Long, Deque<TimeAndDims>>) facts).subMap(timeStart, timeEnd);
-      final Map<Long, Deque<TimeAndDims>> rangeMap = descending ? subMap.descendingMap() : subMap;
-      return concat(rangeMap.values(), descending);
-    }
-
-    private Iterable<TimeAndDims> concat(
-        final Iterable<Deque<TimeAndDims>> iterable,
-        final boolean descending
-    )
-    {
-      return () -> Iterators.concat(
-          Iterators.transform(
-              iterable.iterator(),
-              input -> descending ? input.descendingIterator() : input.iterator()
-          )
-      );
-    }
-
-    @Override
-    public Iterable<TimeAndDims> keySet()
-    {
-      return concat(facts.values(), false);
-    }
-
-    @Override
-    public int putIfAbsent(TimeAndDims key, int rowIndex)
-    {
-      Long time = key.getTimestamp();
-      Deque<TimeAndDims> rows = facts.get(time);
-      if (rows == null) {
-        facts.putIfAbsent(time, new ConcurrentLinkedDeque<>());
-        // in race condition, rows may be put by other thread, so always get latest status from facts
-        rows = facts.get(time);
-      }
-      // setRowIndex() must be called before rows.add() for visibility of rowIndex from concurrent readers.
-      key.setRowIndex(rowIndex);
-      rows.add(key);
-      // always return EMPTY_ROW_INDEX to indicate that we always add new row
-      return TimeAndDims.EMPTY_ROW_INDEX;
-    }
-
-    @Override
-    public void clear()
-    {
-      facts.clear();
-    }
   }
 
   private class LongMetricColumnSelector implements LongColumnSelector
